+++ components/features/course/form/result-form/link-section.tsx
